+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGES
 
+## v1.3.0
+
+### Added
+  - `plot` for broken FE Spaces are now plotted discontinuously to avoid averaging of values at the grid nodes.
+     This can be disabled by passing the kwarg `average_broken_plots = true` to the `plot` call.
+
 ## v1.2.0 June 4, 2025
 
 ### Changed